Fix host_hash related crash when using NCCL_TOPO_FILE and MNNVL
When NCCL_TOPO_FILE is used the system topology comes from xml which
may not have host_hash set in the cpu nodes, or the given host_hash
might be incorrect for the node it is currently running on. This
change ignores the host_hash given in the XML files and recalculates
it after loading an XML file.
BUG 5219094
https://nvbugspro.nvidia.com/bug/5219094