Fix bug in network profiler
The network plugin profiling relies on NCCL passing a NCCL core handle
(opaque to the net plugin) that contains all the profiler information,
including the parent event handle for the network defined events in the
plugin.

This opaque handle is passed back to NCCL by the network plugin when
starting a new event. In the NCCL core profiler callback the opaque
handle is casted back to the original struct ncclProxySubArgs to access
the profiler information.

The parent event handle is stored in the stepEventHandles array inside
the original struct. This is accessed using the profilerSteps counter.

The bug lied in the fact that when NCCL makes a network plugin call,
the network plugin might not callback into NCCL immediately. For example,
in the case of multisend the network plugin will simply return a request
pointer for the isend back to NCCL. NCCL can then alter the counter
(profilerSteps) that the network plugin will later try to use to
reference the parent event handle when starting a new network event,
resulting in the network event to be parented by the wrong proxy event.

The fix here is to not pass the whole ncclProxySubArgs object to the
network plugin but only the pointer to the parent event handle. Since
the NCCL core profiler callback also needs additional information,
like the profilerContext, the eActivation mask and rank we create a
new ncclProxyEventHandle struct that contains both the parent event
handle and a pointer to the ncclProxySubArgs object for the originating
proxy event.