Squash commit for Fix a bug in ncclMemAlloc() when out of memory
The FABRIC handle cuMemCreate() call could fail for other
reasons and the code would not detect/report the error.

Due to the CUDA handle not being initialized it could then
actually continue and return an invalid memory mapping.

BUG 5211959

Fix a bug in ncclMemAlloc() when out of memory

The FABRIC handle cuMemCreate() call could fail for other
reasons and the code would not detect/report the error.

Due to the CUDA handle not being initialized it could then
actually continue and return an invalid memory mapping.

BUG 5211959

See merge request nccl/nccl!887