Switch RAS cleanup back to atexit()
I switched it from atexit() to __attribute__((destructor)) in nccl/nccl!784
because I didn't think that the atexit() handler would be invoked if
libnccl.so was dynamically unloaded (via dlclose(), as at least some
frameworks apparently do).

This turned out to cause problems with NVSHMEM, which dynamically loads
both OpenMPI and NCCL libraries.  OpenMPI's libopen-pal.so inserts an
interceptor of the madvise(2) system call but, at least in the version I
tried (4.0.0), fails to clean up before it gets unloaded, so the next call
to madvise() is guaranteed to cause a crash.  On exit, if OpenMPI ends up
being unloaded before NCCL, we would get a crash when NCCL terminates its
RAS thread (the madvise() call actually comes from glibc's libpthread.so,
which calls it once the thread terminates to clean up that thread's stack).

I have a strong opinion on this OpenMPI "feature" but that's beside the
point -- it's out there in the wild, being used, and we need to deal with
it.  Basically, we need to ensure that the RAS thread is terminated before
OpenMPI is unloaded.  Experiments show that, unlike
__attribute__((destructor)) functions, all atexit() handlers are invoked on
exit before any shared libraries are unloaded.  Even more importantly, it
turns out that since glibc 2.2.3, when a shared library is dynamically
unloaded via dlclose(), any atexit() handlers it registered are invoked at
that point (per atexit(3) man page), so atexit() is in fact safe for us to
use and we don't need to rely on __attribute__((destructor)) for anything.

BUG 5103681

See merge request nccl/nccl!940