Deprecate ncclResetDebugInit
The function will reset the debug system in the case of changes to the
environment variables that control it. It would be better if this was
not exposed to the public, in particular since it will likely be
replaced with a different mechanism in the future.

The function is marked deprecated from nccl.h.in (so the compiler will
warn if it is used) and the function itself is given the default
visibility attribute instead of using the NCCL_API macro. A new
ncclResetDebugInitInternal function has been added to do the actual
reset, with the existing function calling that. This allows test code
to continue to use the reset function without warning during
compilation.