Add string for IB async event IBV_EVENT_GID_CHANGE (18)
On RoCE, when we have a link flap, the NICs will typically be assigned a new GID index.

This will cause an IB async event to happen with code 18 (IBV_EVENT_GID_CHANGE) and NCCL
previously reported that as "unknown event" and now it will report "GID table changed".

BUG 5277908