Tuning NVLS and CollnetDirect AG/RS busBW
AG and RS needs to consider self-send for NVLS, CollnetDirect should not.
So in non-UB case, as long as collnet is available, for AG and RS, we
should always use CollnetDirect instead of NVLS + Collnet. To this end,
this patch set lower intra-node bandwidth for NVLS algo for AG and RS
by a factor of (ppn - 1.0f) / ppn.

At the same time, we lower the Collnet inter-node bandwidth by a factor
of 0.9 which is an empirical setting.