plugin: fix bug in shared plugin library support
We need to dlopen the net library from both tuner and profiler when
these are provided as part of the net library. Since we don't keep a
cross plugin reference counter for the plugin library the first plugin
reaching a local ref counter of 0 would close the plugin library for
everyone, causing crashes if other plugins try to call finalize after.

Similarly, we need to close only the handle for the corresponding plugin
and not all of them.