Squash commit for Always run QA-team-ci at the end of CI
Noticed that the condition to always run the `QA-team-ci` job at the end of the pipeline was missing. This results in the job being skipped unless all jobs in the pipeline are passing like in this pipeline: https://gitlab-master.nvidia.com/nccl/nccl/-/pipelines/29503695

This condition has been updated to match the condition for `summarize_pipeline` which runs regardless of previous job failures.

Always run QA-team-ci at the end

See merge request nccl/nccl!1020