Squash commit for GB200 - reorder the NET returned by `ncclTopoSelectNets`
This MR adds two improvements for the GB200 graph search:

The image below illustrates the proposed change in the NET dev ordering in `ncclTopoSelectNets`:
![GB200_GPU_to_NIC](/uploads/8080f39ee558721ba42f560a53f14a0c/GB200_GPU_to_NIC.png)

BUG 5294160

search: reorder the net devices for MNNVL

Instead of using complex (and buggy for corner cases) formula, we separate the list of localNet to each GPU.

The idea is to reorder the devices withing the list that is returned by ncclTopoSelectNets.
The key is to determine the right stride to be used in order to scatter the NET devices.
If the stride is too small (ex 1), we will use many rails, but risk collision between communicators on the usage of the devices.
If the stride is too large, we reduce the chances of collisions, but reduce the number of rails used for the communication.

The safest and smaller stride is determined by how many NET devices are seen from a single GPU.

As example, consider the case of GB200-NVL8
- GPU/0-0 has NET/0-0, NET/0-1
- GPU/0-1 has NET/0-1, NET/0-0
- GPU/0-2 has NET/0-2, NET/0-3
- GPU/0-3 has NET/0-3, NET/0-2
- GPU/1-0 has NET/1-0, NET/1-1
- GPU/1-1 has NET/1-1, NET/1-0
- GPU/1-2 has NET/1-2, NET/1-3
- GPU/1-3 has NET/1-3, NET/1-2

The stide here would therefore be 2.

If MNNVL_SCATTER_NET_ENABLE=0 (aka, scattering is disabled), we append the list of NET for each of the GPUs and remove the duplicates.
ncclTopoSelectNets will then return the list of all the NETs associated to the node, which then will be considered in order for the graph search.

- split_mask=0x0: NET/0-0, NET/0-1, NET/0-2, NET/0-3, NET/1-0, NET/1-2, ...

- split mask = 0x1 (comm0 = GPU/0-0, GPU/0-2, GPU/1-0, GPU/1-2, and comm1 = GPU/0-1, GPU/0-3, GPU/1-1, GPU/1-3):
comm0: NET/0-0, NET/0-1, NET/0-2, NET/0-3, NET/1-0, NET/1-2, ...
comm1: NET/0-1, NET/0-0, NET/0-3, NET/0-2, NET/1-2, NET/1-0, ...

If creating 4 channels, comm0 and comm1 will therefore compete on the same NETs while not using available NETs.

If MNNVL_SCATTER_NET_ENABLE=1, the NET will be reordered:

- split_mask=0x0: NET/0-0, NET/0-1, NET/0-2, NET/0-3, NET/1-0, NET/1-2, ...

- split mask = 0x1 (comm0 = GPU/0-0, GPU/0-2, GPU/1-0, GPU/1-2, and comm1 = GPU/0-1, GPU/0-3, GPU/1-1, GPU/1-3):
comm0: NET/0-0, NET/0-2, NET/1-0, NET/1-2, NET/0-1, NET/0-3, ...
comm1: NET/0-1, NET/0-3, NET/1-1, NET/1-3, NET/0-0, NET/0-2, ...

If creating 4 channels, comm0 and comm1 will not compete on the same NETs and every net will be used.

See merge request nccl/nccl!1009