topo: fix hostHash for getSystemFromXml
Fix a crash of PXN on MNNVL systems.

The hostHash we passed to ncclTopoGetSystemFromXml was coming from
the peerInfo table, where hostHash is actually a mix of the hostHash
and the communicator hash. This is different than what is in the XML,
which comes from getHostHash(). For that reason, the systemId was
never found and we ended up with all ranks believing they were system
0, hence trying to use the NICs on the first node for PXN operation,
later resulting in an error when trying to map the SHM segment of a
different node.

To fix the issue, we need to pass getHostHash() to
ncclTopoGetSystemFromXml (which is fast as we buffer the answer with
pthread_once).