Fix fp8 reductions on Blackwell with CUDA 12.8
The FP8 reduction code for Blackwell was not being compiled when using
CUDA 12.8 because of inconsistent use of the NCCL_CUDA_ARCH_SPECIFIC /
NCCL_CUDA_ARCH_FAMILY_SPECIFIC macros.  This was causing failures when
attempting to use AllReduce or ReduceScatter with FP8 datatypes.

Update "device/reduce_kernel.h" and "device/symmetric/generate.py" to
match the PTX docs and get them in sync with each other.

Add the testing with all datatypes to a subset of symmetric kernel
invocations in the perf CI.

BUG 5345494

See merge request nccl/nccl!1058