Squash commit for Move nChannelsPerNetPeer into ncclConfig struct
`nChannelsPerNetPeer` is a a tunable env variable to set the number of channels for NIC traffic. There can be conflicting objectives on tuning for applications using A2A as well as P2P traffic such as tensor and pipeline parallelism. This MR makes the variable available at `ncclCommInitRankConfig()` such that it can be set for each communicator independently and addresses https://nvbugspro.nvidia.com/bug/5288641.

Following the other variables in the `ncclConfig` struct the env variable will overwrite the setting in `ncclCommInitRankConfig()`.

Changes are summarizes as:
- Updated `ncclConfig` struct version to align with the new version.
- Added api-tests for nChannelsPerNetPeer in `ncclConfig` struct.
- Adjusted version check for nChannelsPerNetPeer in `parseCommConfig`
  to align with NCCL version 2.27.2.

BUG 5288641

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
  - Introduced a new configuration option to control the number of channels per network peer.
  - Added support for setting this option via environment variables and configuration structures.

- **Bug Fixes**
  - Improved handling and validation of the new configuration parameter during initialization.

- **Tests**
  - Added new test cases to verify the correct behavior of the channels-per-network-peer configuration in various communication scenarios.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

nChannelPerNetPeer from global env variable to communicator local

See merge request nccl/nccl!1065