profiler: fix event emission in NCCL
Previously, profiler events below the root event were emitted
only if the events above were also tracked by the profiler plugin.
This was done to preserve the full event hierarchy. However, with
the recently added API events that are the new top of the event
tree old plugins no longer receive any event from NCCL. The
current behavior clearly no longer matches the new event hierarchy
because of wrong assumptions. This patch remove such assumption and
allows NCCL to emit events if they have been subscribed to by the
plugin, regardless of the tracking of parent events. This means
the new implementation allows the plugin to track only a sub tree
of the event hierarchy or even individual events.