Squash commit for v2.28 Multi-Request Net API
This MR proposes a Multi Request API for NCCLNet.

User of NCCL net plugin can call isend/irecv with a special value in the request argument hinting the plugin that the current send/recv is part of a multi-send/recv so the plugin can optimize the execution of this group of send/recv calls.

The main use-case is PAT AllGather. In the final stages of PAT AllGather, especially in large scale when the communiator size is large (e.g., 256, 512), each rank sends many chunks of data using send/recv to a single peer. The optimization space, in this case, is both utilizing that the fact that the chunks are sent to a single peer and also the fact that these chunks reside in a striding memory pattern. Within the plugin, both of these aspects could be optimized in different ways. The most straight forward optimization is to use a single doorbell for all of the send/recv. Further optimization can be done to take advantage of the striding pattern (the current API proposal does not suggest it but the plugin can auto-detect it if proved efficient) and allow efficient zero-copy (aka UB) even when chunks are small.

After the plugin is called with `NCCL_NET_MULTI_REQUEST`, the first send/recv that is called following the first call without this hint will allow the the plugin to conclude that the group of send/recv is terminated and the plugin will return a request that can be tested/progressed. In this sense, the API resembles in a way the [FI_MORE API in Libfabric](https://ofiwg.github.io/libfabric/v2.0.0/man/fi_msg.3.html).

Related links/bugs/issues:

* https://nvbugspro.nvidia.com/bug/4642355
* https://jirasw.nvidia.com/browse/NCCL-1969
* https://jirasw.nvidia.com/browse/NCCL-1697
* https://gitlab-master.nvidia.com/ramin/ncclnet-plugin/-/tree/v10-multi-op-request-api?ref_type=heads
  * Implementation of the API in NCCLNet Plugin.
* https://gitlab-master.nvidia.com/nccl/nccl/-/merge_requests/1043
  * Implementation of the API in NCCL's internal plugin (`net_ib.cc`)
* [[RFE NCCL] User Buffer Support for PAT](https://nvbugspro.nvidia.com/bug/5022955)

Add Multi Request API

User of NCCL net plugin can call isend/irecv with a special value in the request argument hinting the plugin that the current send/recv is part of a multi-send/recv so the plugin can optimize the execution of the group of send/recv.

After the plugin is called with NCCL_NET_MULTI_REQUEST, the first send/recv that is called following the first call without this hint will allow the the plugin to conclude that the group of send/recv is terminated and the plugin will return a request that can be tested/progressed.

See merge request nccl/nccl!1139