Squash commit for Move libvirt bridge interface to the end of possible interfaces
Network interfaces starting with `virbr` are usually virtual bridges to relay
traffic to containers running on the host. These can usually not be used for
socket connections to remote nodes and ncclCommInit will fail if chosen.

Move libvirt bridge interface to the end of possible interfaces

Network interfaces starting with `virbr` are usually virtual bridges to relay
traffic to containers running on the host. These can usually not be used for
socket connections to remote nodes and ncclCommInit will fail if chosen.

See merge request nccl/nccl!1146