Clean up the diagnostics during plugin loading
Don't print the "Failed to find" messages anymore.  Just print the found
one (including the symbol version).  If no valid symbols are found, print
that the plugin is "unsupported".

Print the full paths of the loaded plugins in the final success/failure
messages, obtained via dlinfo().

Print the values of the NCCL_x_PLUGIN environment variables consistently
with the rest of NCCL code.  Bail out early in case of "none"; don't try to
load it.

Add final confirmation messages on successful loading of profiler and tuner
plugins.

Print the plugin name in error messages if opening failed.

Avoid a terminating period in "Could not find:" messages unless there's a
follow-up sentence, as the period could be confused to be part of the file
name.

Unify the handling of empty libName in openPluginLib().

Clean up the use of INFO subsystems, especially ENV/INIT.

BUG 5309312