Squash commit for Fix stack overflow issues on MNNVL systems
Some (misconfigured!) systems come with a default stack size limit of
"unlimited".  Due to an idiosyncracy of GNU libc (see pthread_create(3)),
this actually results in a _decrease_ of the default pthread stack size to
just 2 MB (!), which is insufficient for the graph search above some 50
MNNVL-connected ranks.

Query the current setting and, if needed, issue diagnostics and change the
default using a GNU-libc-specific API.  We don't want to explicitly pass
the attributes to pthread_create because that option won't be available if
we switch to C++ threads.

BUG 5309620

BUG 5356271

BUG 5420198

BUG 5381348

Fix stack overflow issues on MNNVL systems

Some (misconfigured!) systems come with a default stack size limit of
"unlimited".  Due to an idiosyncracy of GNU libc (see pthread_create(3)),
this actually results in a _decrease_ of the default pthread stack size to
just 2 MB (!), which is insufficient for the graph search above some 50
MNNVL-connected ranks.

Query the current setting and, if needed, issue diagnostics and change the
default using a GNU-libc-specific API.  We don't want to explicitly pass
the attributes to pthread_create because that option won't be available if
we switch to C++ threads.

BUG 5309620
BUG 5356271
BUG 5420198

See merge request nccl/nccl!1158