test: weaken perftest dependency from profiler
The perf tests use exampleProfilerStart and exampleProfilerStop to
instrument perf test with profiler support. This requires the example
profiler library to be linked to the perf test binaries. In some cases
this is not done (e.g., QA DVS builds). Since the profiler symbols not
being available is not a requirement unless the users wants to enable
profiling we can make these "weak" to avoid the build error.

BUG: 5439320