net: fix bug in net_v10 compatibility layer
Since NCCL v2.28 (net_v10) the net plugin is initialized multiple times,
once per communicator instance. Every time the network plugin is
initialized it returns a context handle.

Since the net_v10 expects the network plugin to be initialized only once
across all communicators, the net_v10 compatibility layer uses a ref
counter to make sure init calls following the first are silently dropped
(they are not passed to the net_v10 plugins).

net_v11 introduced another important change. It moved
ncclNetCommConfig_t from connect to init. The way the net_v10 compat
layer handles this is by copying the ncclNetCommConfig_t handle to the
context handle, so that the net_v10 compat layer can pass it to connect
later on. Because of the ref counter bypass for init calls the context
was not being set to the ncclNetCommConfig_t handle, which caused a
segfault when there are multiple threads per processes. In this case
only the first rank calling into net_v10 init gets the context set to
the net config handle, the other ranks get nullptr and segfault.

This patch makes sure that the net_v10 compat layer always sets the
context to the ncclNetCommConfig_t handle.

Fixes BUG: 5449957