topo: uses P2C to detect the GDR level
When the user sets NCCL_GDR_LEVEL=PIX, the old code used to decide to
use GDR over the P2C link as it's considered to be C2C.
This behavior causes hangs on CX8.

This is not correct and should not be the case.
If the GDR level is set to PIX, no GDR, no P2C connection should be used
with GDR.

This commit fixes the behavior, forcing the user to set P2C explicitly
or relying on the default value which is P2C if NCCL_GDR_C2C is set, or
PXB if not.