Fix a device crash on RTX PRO 6000 Blackwell
It appears that the compiler in CUDA versions prior to 13.0 miscompiles
the symmetric AllReduce and ReduceScatter kernels when targetting sm_120
(such as RTX PRO 6000 Blackwell).

The kernels crash in the destructor of class ncclLsaBarrierSession,
seemingly because this->index (which is set to blockIdx.x in the
constructor, and never changed) is inadvertently getting clobbered.

Add just enough redundant code around the crash site to make the
compiler avoid this bug.  The extra code is only added when targetting
sm_120 and when using CTK older than 13.0.

BUG 5448691