graph_test: properly duplicate hosts for NVLDs
The graph_test will now use the topology from a single node and
duplicate it to fake large NVDs.
This is consistent with how NCCL does it.

The size of the NVLD is obtained from the name <NAME>-NVLD<X> will create
an NVLD of X hosts of topology <NAME>.

Note: this commit still support the old way of doing