Squash commit for Device API perf test additional cmd line options
## Summary

Perf test for multime relied on hard coded values for the requirements to launch the kernels.
This was sub-optimal for the CTA count, since it locks in a certain performance.
And it is critical for multimem, since if multimem is required by the kernel it needs to be set, but we cannot set it always because that would fail on non-multimem systems.

## Implementation
Following the practices of the perf tests, 2 additional cmdline args are implemented.

## Testing
Has been tested manually on P-Tyche (multimem) and works correctly there.

Also tested on gpu-comms (non-multimem) it correctly detects that multimem is not available.
It also works correctly on gpu-comms.

I don't think this is included in CI.

## Tracking
Is related to 

https://nvbugspro.nvidia.com/bug/5457969

https://nvbugspro.nvidia.com/bug/5460313

https://nvbugspro.nvidia.com/bug/5468885

manual rebase

See merge request nccl/nccl!1224