Squash commit for Clean up the behavior of symmetric memory API calls when the memory is not supported
## Summary
Fixes a regression that would cause a failure if ncclCommWindowRegister() was invoked on a system that does not support symmetric memory (e.g., when NCCL is run with NCCL_CUMEM_ENABLE=0). Also adds a missing check for symmetric memory support to ncclDevCommCreate().

## Implementation
There was an unmatched ncclGroupEndInternal() in ncclCommWindowRegister(), resulting in a failure. Ensure that we are always within a group by the time the exit path within ncclCommWindowRegister() is taken.

While ncclCommWindowRegister() is meant to silently handle the case of no symmetric memory support, ncclDevCommCreate() WARNs and fails with ncclInvalidUsage.  That's because there's no credible fallback mechanism in case of ncclDevCommCreate().

## Testing
Manual testing. The bug was originally reported by the QA so is presumably part of the QA testing process.

## Tracking
BUG 5470772

Fix the handling of symmetric memory without CUMEM

An umatched ncclGroupEndInternal() would cause a failure if
ncclCommWindowRegister() was invoked on a system that does not support
symmetric memory.

BUG 5470772

See merge request nccl/nccl!1244