net_ib: do not check for data direct on devs not in HCA list
NCCL should not try to find data direct path for devices that are not in
the NCCL_IB_HCA list.

This causes confusing logs to the user and does not results in the
cleanest implementation of CX8 support.

Here, we first make sure that the device is in the HCA list before
checking for data direct path.