plugin: fix tuner plugin logs
When loading the tuner plugin from the network plugin library (i.e.,
when the NCCL_TUNER_PLUGIN is unset) the plugin loading code still tries
to load the default 'libnccl-tuner.so' plugin and reports an error
claiming the internal plugin will be used instead. After that the tuner
from the network plugin library is loaded instead.

This patch removes the confusing message.

BUG: 5458648