improve safety of dev comm create codepaths
Recently, ncclDevCommCreate was reimplemented to support NCCL
groups to enable multi-GPU per process use cases. As a result of
these changes, device communicator creation is not complete until
the end of the scope of the group the call is in. This commit
modifies perf test code paths related to device comm creation to
ensure that groups properly finish before relying on output group
structures.

This change has 2 parts:
 - Split 2 device communicator create attempts (initial multimem
   check and final create attempt) into separate groups
 - Use a temporary host communicator for the initial dev comm
   check to avoid messing up internal state of the original host
   communicator in case the multimem check fails