topo: netdev policy max bounds the number of NET devs
When the user specifies the maximum number of NET devs per GPU, NCCL
shouldn't consider more than the prescribed amount.

We stop adding devices when the threshold is reached.

For most cases, we expect the MAX to be reached right after the
preferred NET addition as it adds the right amount of NET devs per GPU.
If the specified value is higher than the number of NET devs actually
accessible to the GPUs, we should keep adding until we reach the
threshold.

BUG 5195850