Squash commit for Avoid busy loop
## Summary:

Upstreamed from a PR on github by Saif Hasan (Meta). His description:

> Bootstrap operations enable Communicator to initialize. It happens as part of startup of a Job when PG/Communicators are created.
> 
> Bootstrap offers APIs to perform I/O over Sockets (e.g. AllGathers, Barrier, Send/Recv). For large scale jobs the order of bootstrap operations would have higher skew (e.g. late joiners) as well as overall I/O (number of messages & data-size) is large.
> 
> In our large scale training, we observed that Bootstrap chokes CPU at 100% (one core per process) and lasts much longer as job scaled. This competes with the other CPU intensive resources especially during startup phase of the job.
> 
> We found this by tracing CPU for the duration of Job's initialization. And socket.cc being the highest contributor for the same.



## Implementation: Describe the implementation at a high level. Things to cover:

From Saif:

> For non-blocking Socket Read/Writes, the function socketProgress which in-turn calls socketProgressOpt will return if there are no bytes to READ or WRITE. And socketWait will loop the socketProgress again causing the busy read/write loop.
> 
> The fix is to use poll for socketWait, that'll wait until the socket becomes readable / writable to avoid busy loop, and also permit us to abort the operation within 1ms duration. This is similar to what the existing code do for socketPollConnect.

A new parameter, NCCL_SOCKET_POLL_TIMEOUT_MSEC is added which, if set to a value other than the default (0),
will replace spinning with a `poll`. This should reduce CPU usage. This can be useful in real apps to reduce wall clock time.

## Testing

I wrote an apitest to verify that it works, but it cannot be run in CI without improvements to NCCL_PARAM (since the value of parameters is cached and I need to change it to compare the init times). It also requires a `sleep` to be injected in the connection process to observe the results clearly... however, it does make a difference without the sleep so we could make a test if we could change the cached value for the parameter. Since the apitest cannot work currently, but may work in the near future, I've commented out the test.

More details in the comments.

## Tracking

Upstreamed from [Use poll to avoid busy loops for bootstrap socket-io by saifhhasan · Pull Request #1759 · NVIDIA/nccl](https://github.com/NVIDIA/nccl/pull/1759)

Use poll to avoid busy loops for bootstrap socket-io

See merge request nccl/nccl!1153