Squash commit for Performance Regression Testing
## Summary
Adds a nightly performance regression test for NCCL.  This currently is not comparing to any baseline, so it will not fail. However that functionality is planned to be added in the next few weeks.

## Implementation
Adds the `perf-regression-eos` job to .gitlab-ci.yml that only runs on a schedule or web trigger.
The job generates an sbatch script using gcperf-tools for performance testing on EOS.
The slurm job runs the tests and produces a tarball, which gcperf-tools can then convert to a CSV.
Once functional, the resulting CSV will be compared to the previous night's to determine if there's been a regression.

EOS is the only cluster the regression testing runs on in this iteration, however more clusters will be added in the near future.

## Testing
Testing by manually triggering the CI pipeline off of this branch.

## Tracking
NvBug: https://nvbugspro.nvidia.com/bug/5377685

Use gcperf-tools to run performance regression testing

See merge request nccl/nccl!1249