Squash commit for p2p: use a group-sized schedule
## Summary
The MR generalizes the existing p2p scheduling to the case of un-even GPU count per node.

This MR introduces "groups" instead of the nodes, which allows for more flexibility and better scheduling.

It solves two commonly encountered issues when using a2a with large NVLDs:
1. if the NVLDs countain different number of GPUs, the current code will not follow node boundaries and generate poor PxN aggr factor, which reduces performance
2. when using NVLD36, in some configurations (for ex: 16 channels, 20 NVLDs), the work scheduled per channel is uneven: channels will get 40, 44, 48, or 52 peers, which will cost us 7 rounds instead of 6. (see `python3 nccl_channel_simulator.py -r 36 -n 20 -c 16` with `nccl_channel_simulator.py` available [here](https://confluence.nvidia.com/display/GCG/Topology+and+Graph+Search+Files))


## Implementation

Both problems can be solved by dividing the NVLDs into groups of GPUs, each group of equal size (max 8).
Then the groups will be used to schedule work on the channels. avoiding a single group to be spread onto different channels solves the problem number 2. While dividing the nodes into groups solves the problem number 1.

We proceed in 2 steps:
1. find the group size. A node has to contain exactly a number of groups ( = group size must be a divider of the node size). We loop over the nodes and determine the group size as the GCD of all the node sizes.
2. we use the quadratic update formula to visit all the groups, and schedule the ranks linearly within a group.
For more details, see https://confluence.nvidia.com/display/GCG/All-to-All#AlltoAll-P2PScheduling
3. once the groups are created, we used them to schedule work on the channels

## Testing

- verified that NVLD_SIZE=3 with srun -N 4 --segment=4 delivers the aggregation factor of 4, as expected.
- verified bw at 8GB for a2a (default chunksize and nchannels_per_net_peer)

| config | master  | MR |
| ------ | ------ |------ |
| NVL8 - 2 NVLDs       |    85GB/s    | 88GB/s|
| NVL12/4 - 2 NVLDs       |    47GB/s    | 82GB/s

## Tracking

- https://nvbugs/5429630 [NCCL] All-to-all PXN does not always aggregate messages on GB200

p2p: use a group-sized schedule

The goal of the p2p scheduling is to allow a process to send/recv to both far and close peers simulateously.

A given rank will use up to P2P_NCHANNELS simultaneously to communicate.
Consecutive peers will be matched to channels by groups of size min(localRanks, NCCL_MAX_DEV_WORK_P2P_PER_BATCH), see channels.h.
So the p2p schedule needs to make sure that (1) a group of GPUs from the same node will be kept in sequence, and (2) that we alternate between far and close groups.

The existing code considers that a group is always a node, while in this new version, we construct groups of smaller sizes.
We first find the optimal group size, which has to guarantee that each of the node can be divided into multiple groups.
This is done by finding the GCD of all of the node sizes.

Once we have the group size, we go over all the groups and schedule them.
The order in which we schedule the groups is given by the existing quadratic update ((x*x+x)/2 mod N), which guarantees that both far and close groups will be adjacent in the group list.
Inside a given group, there is no need to shuffle as all the peers within a group will be targeted at the same time.

See merge request nccl/nccl!1183