Fix issues with failed NVLS UB registrations
Don't WARN if the supplied buffer is incompatible with multicast (e.g., is
allocated using cudaMalloc()).  This was especially problematic with the
implicit registrations during graph capture.

Query the CU_POINTER_ATTRIBUTE_ALLOWED_HANDLE_TYPES on the supplied
buffers and bail out early if the memory is clearly not compatible.
This avoids the needless creation and tear-down of multicast objects.

Add missing clean-up code in tryRegisterBuffer() if, e.g.,
cuMulticastBindAddr() fails.  We were failing to release the multicast
object, eventually resulting in resource exhaustion.  This required an
addition of a barrier (copied over from nearly identical code in
nvlsAllocateMem()), as otherwise the cleanup code would race with the
UDS proxy thread that was still busy exporting the handles to peers.

BUG 5498967