Squash commit for [CI] Add ccache for builds on IPP6
## Summary
Adds the use of ccache to speed up builds on IPP6 in CI.

## Implementation
Adds an option to build_nccl.sh that enables ccache by linking gcc, g++, and nvcc to the ccache binary and passing the linked paths to make.  The NVCC variable in common.mk has been updated to allow this override. The CI script configures ccache to utilize a Redis database (hosted on IPP6) as a remote backend to share the cache between all nodes. If the remote backend is down or unreachable for any reason, the build will continue on as normal.

## Testing
Tested using docker/make.sh directly on IPP6 and by running CI with the updates.
In testing, the addition of ccache led to a 77% reduction in build times (excluding cloning and archiving steps).

## Tracking
https://jirasw.nvidia.com/browse/SGCOMDEV-197

Add ccache for ipp6

See merge request nccl/nccl!1287