Rename the members in the exchanged metadata info to reflect the usage better
The `rkey` and `addr` which are exchanged are used by the receiver and sender
differently. Sender populates the `addr` with the address of the CTS FIFO and
the receiver populates the `addr` with the address of the Sizes FIFO.

The sizes is not really a FIFO, thus the `fifo` suffix was removed.