Add NOWARN macro and its NCCLCHECK variants
NOWARN abstracts the use of the litte-used ncclDebugNoWarn
warning-suppression capability, addressing some of its inherent problems
such as preserving the old value and restoring it in case of errors.

NCCLCHECKNOWARN and NCCLCHECKGOTONOWARN combine the commonly used
NCCLCHECK/NCCLCHECKGOTO with NOWARN.

Convert all the existing users of ncclDebugNoWarn to the new macros.
Most of the existing users would fail to reset the value to 0 in case of
an error.