Refactor ncclIbLogCompletionWithError function
No need to provide the function with the ncclIbRequest that caused the failure
as it might be that the completion with error was not originated because of
the request that is currently progressed/tested.

Note that the warning format was changed *not* to print details of the request
as the request details are might not be relevant.