Refactor ncclIbCompletionEventProcess function
No need to provide the function with the ncclIbRequest. Although the completion
event has been generated on a device that the request progressed expects an
event on, the completion might not be related to the request progressed.

The actual request that the event belongs to, is retrieved in the function
itself.