Remove duplicated assignment of LKeys in the send request
The same LKeys are assigned in a different loop below the loops that appends
events.

Note that the number of the iterations in the loop that appends events is as
the number of QPs which is larger than the number of devices. The number of
LKeys is as the number of devices so the loop over devices that assigns the
LKeys is more efficient.