Squash commit for Fix new warning from the segfault fix
## Summary
The fix for the segfault that occurs when you compile with 12.9 and run on 12.1 changed two arrays from "extern void* const var[]" to drop the const. Both are initialized to values at the time of declaration. This caused a warning. This change fixes it by removing the unneeded "extern".

## Implementation
This code is generated by the two generate.py scripts. The scripts were modified to remove the "extern" in the generated code. The "extern"s are unnecessary since the files being generated are not header files, but .cc files. Since the arrays are no longer extern, the compiler does not need to warn since there can be no linker conflict.

## Testing
Compiled without the warning. I also ran apitests, all_reduce_perf, and put_signal_ping_pong_gin.

## Tracking
This is related to nvbugs/5339956.

Fix new warning from the segfault fix

See merge request nccl/nccl!1372