Add ID-based matching scheme for receive requests
This change introduces a new matching scheme of receive requests in the data
transfer protocol, which allows more flexible completion semantics of receive
requests and other capabilities like pre-posting WQEs on the receiver side.
The scheme which the net IB uses in the data transfer (whether it's the
old/legacy scheme which was used before this change, or the new one introduced
in this change) is controlled by the environmental variable
`NCCL_IB_RECEIVER_SIDE_MATCHING_SCHEME` which by default assigned to the value
of `BY_INDEX` which means the scheme is according to the old/legacy scheme, and
`BY_ID` which is the new scheme.

In the index-based scheme, the receiver, upon getting a work completion (CQE),
always retrieved/matched the request (`ncclIbRequest`) based on the `wr_id`
that is delivered in the CQE. Since the `wr_id` assigned to the work request
(and returned in the CQE) was the index of the request in the pool of the
requests stored on the communicator (`ncclIbNetCommBase::reqs[]`) which was
calculated using pointers arithmetics, i.e., `wr_id = req - comm->base.reqs`,
it possible to retrieve the receive request using its index. This design
implicitly restricted the receive work request (WQE) posted on the receive
queue of the QP to be completed only by the corresponding send request (that
matches the tag, size etc.) and that the receive work request cannot be used to
complete any other send request.

In the new scheme implemented in this change, upon getting a CQE, the receiver
retrieves/matches the receive request (`ncclIbRequest`) using an ID delivered
in the Immediate Data field in the work completion (`wc.imm_data`). The value
stored in the Immediate Data field in the work completion is assigned by the
sender when posting the RDMA Write with Immediate work request. This means that
the sender must assign the same ID to the RDMA Write with Immediate work
request that would match the ID of the receive request posted by the receiver.

Even though sender must assign the same ID to the send request of the
corresponding receive request, the ID is *not* passed explicitly in the
CTS message from the receiver to the sender. Instead, the ID's value is
derived both by the sender and the receiver from the "slot" number in the CTS
FIFO to which the send and receiver requests belong. Since both receiver and
sender are synchronized on the "slot" number for every request, the ID is
"synchronized" as well, so such matching scheme is possible without explicitly
passing the ID between the receiver and sender.

In the index-based scheme, the "ID" according to which a request was retrieved
was the index of the request in the pool of the requests
(ncclIbNetCommBase::reqs[]). Such design allowed a trivial retrieval of the
request in O(1) time. In the id-based scheme, the ID is derived from the slot
number in the CTS FIFO. Since there is no direct relation between the slot
number in the CTS FIFO and the index of the request in the pool of the
requests, the request cannot be retrieved using the ID of the request as the
index in the pool of requests. Therefore, a trivial retrieval of the request in
the new scheme would be to search all the requests in the pool to find the
request with the matching ID, which means an O(N) retrieval. Instead, this
change overcomes this inefficiency by adding a new array structure in the
receiver communicator which allows to retrieve the request in O(1) time using
the ID as the index in this new array structure.

The new array is a pointer array (array of pointers to requests) and the ID is
used as the index in this array both for insertion and retrieval which allows
O(1) time for both operations. The ID is "converted" to an index using a simple
hash (modulo) on the request's ID. By design, performing such a hash on an ID
of a request results in collision-free, because the number of requests
outstanding on a single slot is limited to one.

It's possible to avoid the addition of this new array structure in the receiver
communicator by passing the ID of the receive requests explicitly to the sender
in the CTS message. Then, the ID's value could still be the index of the
receive request in the pool of requests, which would allow a trivial O(1)
retrieval on the receiver side. However, passing the ID explicitly in the CTS
message would increase the size of the CTS message. Such increase would
probably require to change how CTS message is delivered, to a "header", which
will hold all the common fields of the CTS message that are the same for all
the receive requests within a multi-receive request and a "body" which would
hold all the data specific to each receive request, so it would be possible to
avoid storing the ID of each receive request in the CTS message for every
receive request in a multi-receive request.

In addition to the new pointer array added to the receiver communicator, the
`ncclIbRequest` structure was also added with a new member to store the ID of
request. Since the ID should fit in the Immediate Data field, the member that
stores the ID of the request is in the `ncclIbRequest` structure was defined to
be of width 32 bits. Note that it's possible to avoid the addition of a new
member ID to the `ncclIbRequest` structure (and compute it whenever needed).
Moreover, this ID is not used when the Immediate Data is not carrying it.
Still, this member can also help during debug for an identification of the
request.

In the index-based matching scheme, in some cases the Immediate Data was used
to carry the size of the send request. This usage, in some cases, is redundant
and this redundancy is what allows reusing the Immediate Data to carry the request
ID.

The relevant case where Immediate Data carried the size of the send request was
only in the case of single send/receive (`nreqs==1`), but in this case, if the
send request is sent using a single RDMA Write with Immediate message, the size
of the send request can be extracted from the work completion itself (from the
`struct ibv_wc::byte_len` field). If the sender decides to send the message
using RDMA Write followed by RDMA Write with Immediate (for example to avoid
the fencing of the data on the requestor side when Adaptive Routing is
enabled), the receiver should get the size of the send request from the
completions records array. Since the receiver does not know in advance the size
of the send request (as it might be smaller than the receive request), the
receiver cannot differentiate between the two protocols that the sender may
follow. To overcome this, the receiver should always check whether the
completions records array was written by the sender before deciding from where
to extract the information about the send size.

Another aspect that is covered by this change is how the size of the send
request extracted. In the most trivial case, when the ID-based scheme is used,
the receiver extracts the size of the send request from the `byte_len` field of
the work completion. The problem that needed to be addressed is what happens in
case a data transfer of a send was done over multiple QPs. When using multiple
QP, multiple work completions are generated on the receiver side and each work
completion holds in the `byte_len` field only the size of the data that the
corresponding QP delivered. This change adds a new member field in the
`ncclIbRequest` structure to allow aggregation of all the data sizes delivered
by all the QPs used for the data transfer.

Note that in index-based scheme, the size that was carried in the Immediate
Data was the size of the *whole* send request, regardless if multiple QPs were
used to transfer the data or not, therefore no aggregation was required on the
receiver side.